+++ app/log_group_retainer/main.tf
@@ -6,6 +6,7 @@ module "lambda" {
   function_source_package_runtime_default = "python3.13"
   function_timeout_seconds_default        = var.function_timeout_seconds_default
   iam_data                                = var.iam_data
+  monitor_data                            = var.monitor_data
   std_map                                 = var.std_map
   vpc_az_key_list_default                 = var.vpc_az_key_list_default
   vpc_data_map                            = var.vpc_data_map
@@ -14,19 +15,12 @@ module "lambda" {
   vpc_segment_key_default                 = var.vpc_segment_key_default
 }
 
-module "lambda_permission" {
-  source                       = "../../lambda/permission"
-  permission_map               = local.create_lambda_permission_map
-  permission_principal_default = "events.amazonaws.com"
-  std_map                      = var.std_map
-}
-
 module "trigger" {
   source                              = "../../event/trigger/base"
-  event_map                           = local.create_trigger_map
-  event_schedule_expression_default   = var.event_schedule_expression_default
+  event_map                           = local.create_trigger_x_map
   event_target_service_default        = "lambda"
   iam_data                            = var.iam_data
+  monitor_data                        = var.monitor_data
   std_map                             = var.std_map
   vpc_az_key_list_default             = var.vpc_az_key_list_default
   vpc_data_map                        = var.vpc_data_map