+++ documentation/example/app/s3_proc/main.tf
@@ -65,8 +65,9 @@ module "s3_trigger" {
       target_service            = "batch"
     }
   }
-  iam_data = data.terraform_remote_state.iam.outputs.data
-  std_map  = module.com_lib.std_map
+  iam_data     = data.terraform_remote_state.iam.outputs.data
+  monitor_data = data.terraform_remote_state.monitor.outputs.data
+  std_map      = module.com_lib.std_map
 }
 
 module "local_config" {