+++ event/trigger/base/main.tf
@@ -10,13 +10,14 @@ resource "aws_cloudwatch_event_rule" "this_rule" {
 
 module "dead_letter_queue" {
   source                          = "../../../sqs/dead_letter_queue"
-  queue_map                       = local.lx_map
+  monitor_data                    = var.monitor_data
   name_append_default             = var.name_append_default
   name_include_app_fields_default = var.name_include_app_fields_default
   name_infix_default              = var.name_infix_default
   name_prefix_default             = var.name_prefix_default
   name_prepend_default            = var.name_prepend_default
   name_suffix_default             = var.name_suffix_default
+  queue_map                       = local.lx_map
   std_map                         = var.std_map
 }